+++ src/realm/query.cpp
@@ -63,18 +63,22 @@ void Query::create()
         fetch_descriptor();
 }
 
-Query::Query(const Query& copy)
+Query::Query(const Query& source): error_code(source.error_code), m_groups(source.m_groups),
+    m_current_descriptor(source.m_current_descriptor), m_table(source.m_table)
 {
-    m_table = copy.m_table;
-    m_groups = copy.m_groups;
-    error_code = copy.error_code;
-    m_view = copy.m_view;
-    m_source_link_view = copy.m_source_link_view;
-    m_current_descriptor = copy.m_current_descriptor;
+    if (source.m_owned_source_table_view) {
+        m_owned_source_table_view = source.m_owned_source_table_view->clone();
+        m_source_table_view = m_owned_source_table_view.get();
+        m_view = m_source_table_view;
+    }
+    else {
+        // FIXME: The lifetime of `m_source_table_view` may be tied to that of `source`, which can easily
+        // turn `m_source_table_view` into a dangling reference.
+        m_source_table_view = source.m_source_table_view;
 
-    // FIXME: The lifetime of `m_source_table_view` may be tied to that of `copy`, which can easily
-    // turn `m_source_table_view` into a dangling reference.
-    m_source_table_view = copy.m_source_table_view;
+        m_source_link_view = source.m_source_link_view;
+        m_view = source.m_view;
+    }
 }
 
 Query& Query::operator = (const Query& source)
@@ -82,13 +86,23 @@ Query& Query::operator = (const Query& source)
     if (this != &source) {
         m_groups = source.m_groups;
         m_table = source.m_table;
-        m_view = source.m_view;
-        m_source_link_view = source.m_source_link_view;
 
-        // FIXME: The lifetime of `m_source_table_view` may be tied to that of `source`, which can easily
-        // turn `m_source_table_view` into a dangling reference.
-        m_source_table_view = source.m_source_table_view;
-        m_owned_source_table_view = nullptr;
+        if (source.m_owned_source_table_view) {
+            m_owned_source_table_view = source.m_owned_source_table_view->clone();
+            m_source_table_view = m_owned_source_table_view.get();
+            m_view = m_source_table_view;
+
+            m_source_link_view.reset();
+        }
+        else {
+            // FIXME: The lifetime of `m_source_table_view` may be tied to that of `source`, which can easily
+            // turn `m_source_table_view` into a dangling reference.
+            m_source_table_view = source.m_source_table_view;
+            m_owned_source_table_view = nullptr;
+
+            m_source_link_view = source.m_source_link_view;
+            m_view = source.m_view;
+        }
 
         if (m_table)
             fetch_descriptor();