+++ test/test_query.cpp
@@ -8565,4 +8565,34 @@ TEST(Query_MoveDoesntDoubleDelete)
         q2 = std::move(q1);
     }
 }
+
+// Ensure that coyping a Query copies a restricting TableView if the query owns the view.
+TEST(Query_CopyRestrictingTableViewWhenOwned)
+{
+    Table table;
+
+    {
+        Query q1(table, std::unique_ptr<TableViewBase>(new TableView()));
+        Query q2(q1);
+
+        // Reset the source query, destroying the original TableView.
+        q1 = {};
+
+        // Operations on the copied query that touch the restricting view should not crash.
+        CHECK_EQUAL(0, q2.count());
+    }
+
+    {
+        Query q1(table, std::unique_ptr<TableViewBase>(new TableView()));
+        Query q2;
+        q2 = q1;
+
+        // Reset the source query, destroying the original TableView.
+        q1 = {};
+
+        // Operations on the copied query that touch the restricting view should not crash.
+        CHECK_EQUAL(0, q2.count());
+    }
+}
+
 #endif // TEST_QUERY